pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cardiomas"
dynamic = ["version"]
description = "CardioMAS — Fresh Agentic RAG runtime for grounded dataset understanding"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Vajira Thambawita", email = "vajira@simula.no"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["agentic-rag", "dataset", "retrieval", "grounded-qa", "ecg"]

dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "pypdf>=4.0",
    "rank-bm25>=0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
cardiomas = "cardiomas.cli.main:app"

[project.urls]
Homepage = "https://github.com/vlbthambawita/CardioMAS"
Repository = "https://github.com/vlbthambawita/CardioMAS"
Issues = "https://github.com/vlbthambawita/CardioMAS/issues"
"HuggingFace Dataset" = "https://huggingface.co/datasets/vlbthambawita/ECGBench"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
# Only match vX.Y.Z release tags — ignore pre-release tags like v0.3.0.dev1
# which cannot be bumped by the guess-next-dev scheme.
tag_regex = "^v(?P<version>[0-9]+\\.[0-9]+\\.[0-9]+)$"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
